Design a Home page

+++ app/controllers/likes_controller.rb
@@ -2,11 +2,16 @@ class LikesController < ApplicationController
   before_action :set_post
 
   def create
+    if @post.user == current_user
+      flash[:alert] = "You cannot like your own post."
+      redirect_to posts_path(@post) and return
+    end
+
     @like = @post.likes.build(user: current_user)
     if @like.save
       respond_to do |format|
         format.turbo_stream
-        format.html { redirect_to posts_path, notice: "User like the post!" }
+        format.html { redirect_to posts_path, notice: "You liked the post!" }
       end
     end
   end
@@ -15,10 +20,15 @@ class LikesController < ApplicationController
   def destroy
     @like = Like.find(params[:id])
     @post = @like.post
+    if @post.user == current_user
+      flash[:alert] = "You cannot unlike your own post."
+      redirect_to posts_path(@post) and return
+    end
+
     if @like.destroy
       respond_to do |format|
         format.turbo_stream
-        format.html { redirect_to posts_path, notice: "User Unlike the post!" }
+        format.html { redirect_to posts_path, notice: "You Unliked the post!" }
       end
     end
   end
